Split out systemd-boot
This is another take on splitting the systemd userspace
and efi utilities into a standalone package. Putting these all into
a standalone package makes sense for a couple reasons. First
they are largely redundant on a grub2 based system and just waste
disk space. Secondly, in the case of a systemd-boot system having
them in a dedicated package may be useful if the systemd-boot
efi utility needs to be signed/released on a slightly different
schedule from the main systemd packages. It also allows adjusting
package dependencies/etc as needed when grub isn't installed.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>