Import newer commit of 256-stable instead of backporting patches
Instead of backporting patches in the rpm repository, let's import
a newer commit of the 256-stable branch instead. This makes it easier
for me to sync changes from the Fedora spec to the CentOS Hyperscale
spec as I don't have to mess around with excluding patches if I want
to move to a different stable commit. It also cleans up the rpm repository
as we need a single line modification to import new fixes instead of
having to spend time generating patches which clutter the rpm repository.